gradient-clipped energy
