+++ bgflow/distribution/energy/__init__.py
@@ -45,7 +45,7 @@ Particle Box
     HarmonicParticles
 
 ===============================================================================
-Linlogcut
+Clipped Energies
 ===============================================================================
 
 .. autosummary::
@@ -53,6 +53,7 @@ Linlogcut
     :template: class.rst
 
     LinLogCutEnergy
+    GradientClippedEnergy
 
 ===============================================================================
 Base
@@ -73,7 +74,7 @@ from .particles import *
 from .lennard_jones import *
 from .openmm import *
 from .multi_double_well_potential import *
-from .linlogcut import *
+from .clipped import *
 from .openmm import *
 from .xtb import *
 from .ase import *
