+++ bgflow/distribution/energy/clipped.py
@@ -1,7 +1,8 @@
-from ...utils.train import linlogcut
+from ...utils.train import linlogcut, ClipGradient
 from .base import Energy
 
-__all__ = ["LinLogCutEnergy"]
+
+__all__ = ["LinLogCutEnergy", "GradientClippedEnergy"]
 
 
 class LinLogCutEnergy(Energy):
@@ -24,3 +25,14 @@ class LinLogCutEnergy(Energy):
     def _energy(self, *xs, **kwargs):
         u = self.delegate.energy(*xs, **kwargs)
         return linlogcut(u, high_val=self.high_energy, max_val=self.max_energy)
+
+
+class GradientClippedEnergy(Energy):
+    """An Energy with clipped gradients. See `ClipGradient` for details."""
+    def __init__(self, energy: Energy, gradient_clipping: ClipGradient):
+        super().__init__(energy.event_shapes)
+        self.delegate = energy
+        self.clipping = gradient_clipping
+
+    def _energy(self, *xs, **kwargs):
+        return self.delegate.energy(*((self.clipping(x) for x in xs)), **kwargs)
